🎨 Palette: UX improvements and CI pipeline fixes
UX Improvements:
- Added "Skip to content" link for keyboard accessibility.
- Replaced Link with NavLink for semantic navigation and ARIA support.
- Implemented visual active states (bottom border) for better color accessibility.
- Fixed unescaped apostrophes in UI components.
- Moved useToast to a separate hook file to resolve Fast Refresh warnings.

CI & Code Quality Fixes:
- Fixed multiple linting errors across the codebase (global definitions, unused variables, any types, regex escapes).
- Configured 'Security Scan' job to continue-on-error when SNYK_TOKEN is missing.
- Synchronized package-lock.json.
- Removed junk workspace artifact.
